requirements.txt
kivy
pyjnius
webview-android
pywebview[gtk] # Using GTK for Linux, pywebview handles macOS/Windows automatically
watchdog
